move things around a bit
