+++ src/parse/index.js
@@ -1,7 +1,7 @@
 const redirects = require('./page/redirects');
 const disambig = require('./page/disambig');
 const preProcess = require('./preProcess');
-const parse_sections = require('./sections');
+const parse_sections = require('./section');
 
 //convert wikiscript markup lang to json
 const main = function(wiki) {
@@ -26,7 +26,7 @@ const main = function(wiki) {
   //give ourselves a little head-start
   wiki = preProcess(wiki);
   //parse all the headings, and their texts/sentences
-  r.sections = parse_sections(r, wiki);
+  r.sections = parse_sections(wiki);
 
   // get categories from last section
   // wiki = parse_categories(r, wiki);
