+++ src/parse/section/recursive/image.js
