+++ src/parse/section/sentence/line.js
@@ -1,20 +1,20 @@
-const helpers = require("../../lib/helpers");
-const parse_links = require("./links");
-const i18n = require("../../data/i18n");
-const cat_reg = new RegExp("\\[\\[:?(" + i18n.categories.join("|") + "):[^\\]\\]]{2,80}\\]\\]", "gi");
+const helpers = require('../../../lib/helpers');
+const parse_links = require('./links');
+const i18n = require('../../../data/i18n');
+const cat_reg = new RegExp('\\[\\[:?(' + i18n.categories.join('|') + '):[^\\]\\]]{2,80}\\]\\]', 'gi');
 
 //return only rendered text of wiki links
 const resolve_links = function(line) {
   // categories, images, files
-  line = line.replace(cat_reg, "");
+  line = line.replace(cat_reg, '');
   // [[Common links]]
-  line = line.replace(/\[\[:?([^|]{2,80}?)\]\](\w{0,5})/g, "$1$2");
+  line = line.replace(/\[\[:?([^|]{2,80}?)\]\](\w{0,5})/g, '$1$2');
   // [[File:with|Size]]
-  line = line.replace(/\[\[File:?(.{2,80}?)\|([^\]]+?)\]\](\w{0,5})/g, "$1");
+  line = line.replace(/\[\[File:?(.{2,80}?)\|([^\]]+?)\]\](\w{0,5})/g, '$1');
   // [[Replaced|Links]]
-  line = line.replace(/\[\[:?(.{2,80}?)\|([^\]]+?)\]\](\w{0,5})/g, "$2$3");
+  line = line.replace(/\[\[:?(.{2,80}?)\|([^\]]+?)\]\](\w{0,5})/g, '$2$3');
   // External links
-  line = line.replace(/\[(https?|news|ftp|mailto|gopher|irc):\/\/[^\]\| ]{4,1500}([\| ].*?)?\]/g, "$2");
+  line = line.replace(/\[(https?|news|ftp|mailto|gopher|irc):\/\/[^\]\| ]{4,1500}([\| ].*?)?\]/g, '$2');
   return line;
 };
 // console.log(resolve_links("[http://www.whistler.ca www.whistler.ca]"))
